sonarlint

+++ src/main/java/de/fs92/defi/contractutil/Approval.java
@@ -40,7 +40,7 @@ public class Approval {
       String address = toAllowContract.getAddress();
       Wad18 allowance = new Wad18(
               contract.allowance(contractNeedsProvider.getCredentials().getAddress(), address).send());
-      if (allowance.compareTo(MINIMUM_APPROVAL_ALLOWANCE) < 0) {
+      if (allowance.toBigInteger().compareTo(MINIMUM_APPROVAL_ALLOWANCE) < 0) {
         logger.warn("DAI ALLOWANCE IS TOO LOW {}", allowance);
         approve(address, toAllowContract.getClass().getName());
       }
